feat: add the ability to edit a quote
Add the ability to edit a quote by splitting quote creation into two scenes:
the first scene, 'quote shaping,' gathers the main inputs (text, author, origin),
and the second scene allows deciding whether to confirm and post the quote
or to edit selected aspects of the quote in the future.
Refactor code for the below-mentioned requirement.